ATR-838: renamied folder, namespace and analyser
